refacto-webapp: extract signup form from signup page
